# [PERF] stock,mrp: optimize _compute_quantities