The performance when searching by `qty_available`, `virtual_available`, `free_qty`, `incoming_qty`, or `outgoing_qty` on databases with a large number of `product.product`s is currently poor. This is part of a two-part fix with #224031

Both `product.product.type` and `product.product.uom_id` are computed from `product_tmpl_id`. With large databases, `_compute_related()` is a large bottleneck when running `_compute_quantities()`

We fix this performance issue by directly fetching the fields instead.

Benchmark (with #224031 applied)
| `product.product` count | Before      | After      |
| ----------------------- | ----------- | ---------- |
| 500,000                 | 144s        | 78s        |

opw-4930856