@juwu-odoo 2025-09-08
Reworked in https://github.com/odoo/odoo/pull/225826